Bindings: finish configured binding cleanup
